警告が出ないように修正

+++ src/alphazero_learn.cpp
@@ -82,7 +82,7 @@ void alphaZero() {
 
     //データを取得
     std::vector<LearningData> validation_data = loadData(validation_kifu_path);
-    assert(validation_data.size() >= validation_size);
+    assert(validation_data.size() >= (uint64_t)validation_size);
 
     //データをシャッフルして必要量以外を削除
     std::mt19937_64 engine(0);
